pyproject.toml
[project]
name = "snowflake-dbt-quality-ref"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-autofix>=0.18.7",
    "dbt-core>=1.11.3",
    "dbt-snowflake>=1.11.1",
    # Added to help with local dev. Not needed for dbt runtime.
    "apache-airflow>=3.1.7",
    "apache-airflow-providers-snowflake>=6.9.0",
    "astronomer-cosmos>=1.13.0",
]
